refactor: 공고 리스트 커스텀 훅 리팩터링
